changed api resource to controllers for create and edit
