+++ app/Http/Controllers/ReviewController.php
@@ -1,55 +1,19 @@
 <?php
 
 namespace App\Http\Controllers;
+use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use App\Comment;
 use App\Review;
 use Illuminate\Http\Request;
 class ReviewController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {       
-        return view('review-vue.review');
-    }
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
+    public function create(\App\Review $review)
     {
-        $review = new Review();
-        $review->reviews = $request->input('reviewCreation'); 
-        $review->save();
+        return view('review-vue.create');
     }
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
+    public function edit(\App\Review $review)
     {
-        $updateReview = Review::find($id);
-        $updateReview->reviews = $request->input('reviewsEdits');
-        $updateReview->save();
-    }
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        $review = Review::find($id);
-        $review->delete();
+        return view('review-vue.reviewEdit')->withReview($review);
     }
 }
