+++ resources/js/components/CommentsComponent.vue
@@ -3,7 +3,7 @@
         <div class="col comments-card_single-comments__wrapper">
             <div class="card-body comment-card__single-card" >
                {{this.$props.comment.comments}}
-                <a class="review-card__single-review_edit-comment right-side"  :href="`/api/comments/${$props.comment.id}/edit`">🖉</a>
+                <a class="review-card__single-review_edit-comment right-side"  :href="`/comments/${$props.comment.id}/edit`">🖉</a>
                 <form  class="review-card__single-review_remove-comment right-side" @submit="submit($props.comment.id)">
                     <input :disabled="isDisable" type="submit" class="button-remove p-r" name="btnDelete" value="⨯">
                 </form>
