+++ routes/api.php
@@ -20,23 +20,7 @@ Route::get('/cake-component/review', function(){
 Route::get('/cake-component/comments/{review}', 'CommentController@fetchComments');
 
 Route::apiResources([
-    '/comments' => 'Api/CommentsController',
+    '/comments' => 'CommentController',
     '/reviews' => 'ReviewController',
 ]);
 
-Route::get('/reviews/create', function (\App\Review $review)
-{
-    return View::make('review-vue.create');
-});
-Route::get('/comments/create/{review}', function (\App\Review $review)
-{
-    return View::make('review-vue.commentCreate')->withReviewId($review->id);
-});
-Route::get('/comments/{comment}/edit', function (\App\Comment $comment)
-{
-    return view('comments.edit')->withCommentId($comment->id);
-});
-Route::get('/reviews/{review}/edit', function(\App\Review $review)
-{
-    return view('review-vue.reviewEdit')->withReview($review);
-});
